--- a/QCScaling/Project.toml
+++ b/QCScaling/Project.toml
@@ -5,9 +5,11 @@ version = "0.1.0"
 
 [deps]
 BenchmarkTools = "6e4b80f9-dd63-53aa-95a3-0cdb28fa8baf"
+OhMyThreads = "67456a42-1dca-4109-a031-0a68de7e3ad5"
 ProgressBars = "49802e3a-d2f1-5c88-81d8-b72133a6f568"
 StatProfilerHTML = "a8a75453-ed82-57c9-9e16-4cd1196ecbf5"
 Statistics = "10745b16-79ce-11e8-11f9-7d13ad32a3b2"
 
 [compat]
+OhMyThreads = "0.8.3"
 Statistics = "1.11.1"