--- a/QCScaling/src/contexts.jl
+++ b/QCScaling/src/contexts.jl
@@ -3,14 +3,14 @@ struct Context
     parity::Int
     function Context(pos, parity)
         nqubit = length(first(pos))
-        @assert parity in [0, 1]
+        @assert parity in (0, 1)
         @assert length(pos)==(2^(nqubit -1) + 1) "Not the right number of POs"
         return new(pos, parity)
     end
 end
 
 function Context(generator::ParityOperator, base_cxt::Context)
-    pos = [generator + po for po in base_cxt.pos]
+    pos = Ref(generator) .+ base_cxt.pos
     return Context(pos, base_cxt.parity)
 end