## main
```julia
julia> @benchmark QCScaling.score($states, $rep, $goal, $cxt_master)
BenchmarkTools.Trial: 324 samples with 1 evaluation per sample.
 Range (min … max):  12.957 ms … 60.440 ms  ┊ GC (min … max):  0.00% … 11.38%
 Time  (median):     15.207 ms              ┊ GC (median):    12.73%
 Time  (mean ± σ):   15.409 ms ±  3.348 ms  ┊ GC (mean ± σ):  13.98% ±  5.13%

         ▄█       ▂▆
  ▃▂▃▁▂▂▅███▆▅▃▄▆████▃▃▃▂▃▂▃▁▂▁▂▁▁▂▁▃▁▂▁▁▁▁▂▁▂▁▁▁▁▁▁▁▁▁▁▁▁▂▁▃ ▃
  13 ms           Histogram: frequency by time        22.2 ms <

 Memory estimate: 44.57 MiB, allocs estimate: 515345.
```


## this PR
```julia
julia> @benchmark QCScaling.score($states, $rep, $goal, $cxt_master; multi_threading=false)
BenchmarkTools.Trial: 1657 samples with 1 evaluation per sample.
 Range (min … max):  2.471 ms …  17.251 ms  ┊ GC (min … max):  0.00% … 81.34%
 Time  (median):     2.553 ms               ┊ GC (median):     0.00%
 Time  (mean ± σ):   3.016 ms ± 977.294 μs  ┊ GC (mean ± σ):  15.02% ± 18.20%

  ▆█▅▁                                  ▁▃▄▃▁
  ████▅▅▅▅██▆▆▄▃▃▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▇█████▇▇▆▅▅▃▃▃▁▃▁▁▁▁▁▃ █
  2.47 ms      Histogram: log(frequency) by time      5.33 ms <

 Memory estimate: 10.14 MiB, allocs estimate: 152663.
```

this is about 5x faster than before.

And turning on multi_threading gives another 2-3x faster using 4 cores